Extract ColumnReader interface and ColumnReaderFactory in Parquet
