+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/HiveTestUtils.java
@@ -20,6 +20,7 @@ import io.airlift.units.DataSize;
 import io.trino.PagesIndexPageSorter;
 import io.trino.metadata.Metadata;
 import io.trino.operator.PagesIndex;
+import io.trino.parquet.reader.ColumnReaderFactory;
 import io.trino.parquet.reader.ParquetReader.ParquetReaderFactory;
 import io.trino.plugin.hive.authentication.NoHdfsAuthentication;
 import io.trino.plugin.hive.azure.HiveAzureConfig;
@@ -139,7 +140,7 @@ public final class HiveTestUtils
         return ImmutableSet.<HivePageSourceFactory>builder()
                 .add(new RcFilePageSourceFactory(TYPE_MANAGER, hdfsEnvironment, stats, hiveConfig))
                 .add(new OrcPageSourceFactory(new OrcReaderConfig(), hdfsEnvironment, stats, hiveConfig))
-                .add(new ParquetPageSourceFactory(hdfsEnvironment, stats, new ParquetReaderFactory(), new ParquetReaderConfig(), hiveConfig))
+                .add(new ParquetPageSourceFactory(hdfsEnvironment, stats, new ParquetReaderFactory(new ColumnReaderFactory()), new ParquetReaderConfig(), hiveConfig))
                 .build();
     }
